ref(onboarding): Split Cordova onboarding doc into multiple files
